+++ dotnet/Services/CybersourceApi.cs
@@ -167,7 +167,7 @@ namespace Cybersource.Services
 
                 Console.WriteLine($"- SendReportRequest: [{response.StatusCode}] {responseContent} -");
 
-                //_context.Vtex.Logger.Debug("SendReportRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\nv-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}\n[{response.StatusCode}]\n{responseContent}");
+                _context.Vtex.Logger.Debug("SendReportRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\nv-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}\n[{response.StatusCode}]\n{responseContent}");
             }
             catch (Exception ex)
             {
@@ -218,6 +218,15 @@ namespace Cybersource.Services
                         ProxyTokenResponse proxyToken = JsonConvert.DeserializeObject<ProxyTokenResponse>(proxyTokenSendResponse.Message);
                         digest = proxyToken.Tokens[0].Placeholder;
                     }
+                    else
+                    {
+                        Console.WriteLine("Did not calculate digest!");
+                    }
+
+                    /// TESTING ------------------------------------------
+                    //Console.WriteLine("     !!!!    OVERRIDING CARD DATA FOR TESTING    !!!!    ");
+                    //digest = await this.GenerateDigest(jsonSerializedData);
+                    /// TESTING ------------------------------------------
 
                     request.Headers.Add($"{CybersourceConstants.PROXY_HEADER_PREFIX}Digest", digest); // Do not pass this header field for GET requests. It is a hash of the JSON payload made using a SHA-256 hashing algorithm.
                 }
@@ -246,22 +255,22 @@ namespace Cybersource.Services
                 };
 
                 Console.WriteLine($"- SendRequest: [{response.StatusCode}] - ");
-                //StringBuilder sb = new StringBuilder();
-                //foreach (var header in request.Headers)
-                //{
-                //    string headerName = header.Key;
-                //    string headerContent = string.Join(",", header.Value.ToArray());
-                //    sb.AppendLine($"{headerName} : {headerContent}");
-                //    //Console.WriteLine($" |-| {headerName} : {headerContent}");
-                //}
+                StringBuilder sb = new StringBuilder();
+                foreach (var header in request.Headers)
+                {
+                    string headerName = header.Key;
+                    string headerContent = string.Join(",", header.Value.ToArray());
+                    sb.AppendLine($"{headerName} : {headerContent}");
+                    //Console.WriteLine($" |-| {headerName} : {headerContent}");
+                }
 
-                //_context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{sb}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
-                //_context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
-                _context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n[{response.StatusCode}]\n{responseContent}");
+                _context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n{sb}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
+                //_context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
+                //_context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n[{response.StatusCode}]\n{responseContent}");
             }
             catch (Exception ex)
             {
-                _context.Vtex.Logger.Error("SendRequest", null, $"Error ", ex);
+                _context.Vtex.Logger.Error("SendRequest", "Proxy", $"Error ", ex);
             }
 
             return sendResponse;
@@ -332,10 +341,22 @@ namespace Cybersource.Services
         public async Task<PaymentsResponse> ProcessPayment(Payments payments, string proxyUrl, string proxyTokensUrl)
         {
             PaymentsResponse paymentsResponse = null;
+            /// TESTING ------------------------------------------
+            //Console.WriteLine("     !!!!    OVERRIDING CARD DATA FOR TESTING    !!!!    ");
+            //payments.paymentInformation.card.number = "4111111111111111";
+            //payments.paymentInformation.card.securityCode = "111";
+            //payments.orderInformation.amountDetails.totalAmount = "401";
+            //payments.orderInformation.billTo.postalCode = "28650";
+            /// TESTING ------------------------------------------
+
             string json = JsonConvert.SerializeObject(payments);
             string endpoint = $"{CybersourceConstants.PAYMENTS}payments";
             SendResponse response = await this.SendProxyRequest(HttpMethod.Post, endpoint, json, proxyUrl, proxyTokensUrl);
+            
+
             //SendResponse response = await this.SendRequest(HttpMethod.Post, endpoint, json);
+            //response = await this.SendRequest(HttpMethod.Post, endpoint, json);
+            
             if (response != null)
             {
                 if (response.Success)