+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Conversion Report
 - Decision Manager
 
 ## [0.1.0] - 2021-09-24